feat : core app added and open markdown globally
