+++ static/css/markdown-core.css
@@ -1,0 +1,150 @@
+/* --- DevOpsVaultX Premium Markdown Styling (Final Clean Fix) --- */
+
+.markdown-body {
+    color: #cbd5e1;
+    font-family: 'Inter', sans-serif;
+    line-height: 1.8;
+}
+
+/* Visibility Fix for First Line (H1 / Quotes) */
+.markdown-body > *:first-child {
+    margin-top: 0 !important;
+    padding-top: 5px !important;
+}
+
+.markdown-body h1 {
+    color: #f8fafc !important;
+    font-size: 1.8rem;
+    margin-bottom: 1.2rem;
+    font-weight: 700;
+    display: block !important;
+}
+
+.markdown-body h2 { 
+    font-size: 1.5rem; 
+    border-bottom: 1px solid rgba(255,255,255,0.08); 
+    padding-bottom: 8px;
+    color: #f8fafc;
+}
+
+/* Lists & Bullets */
+.markdown-body ul, .markdown-body ol {
+    list-style-type: none;
+    padding-left: 0;
+    margin: 20px 0;
+}
+
+.markdown-body li {
+    margin-bottom: 12px;
+    padding-left: 30px;
+    position: relative;
+    display: block;
+}
+
+.markdown-body li::before {
+    content: "•";
+    color: #f97316; /* Brand Orange */
+    font-weight: bold;
+    position: absolute;
+    left: 8px;
+    font-size: 1.4rem;
+    line-height: 1.2;
+}
+
+/* Blockquote Fix (Note/SRE Mantra Box) */
+.markdown-body blockquote {
+    background: rgba(249, 115, 22, 0.08) !important;
+    border-left: 4px solid #f97316 !important;
+    padding: 15px 25px !important;
+    margin: 25px 0 !important;
+    border-radius: 8px;
+    font-style: italic;
+    color: #e2e8f0 !important;
+    display: block !important;
+}
+
+/* --- Code Blocks with Copy Button Integration --- */
+.markdown-body pre {
+    position: relative; /* Essential for aligning the copy button */
+    background: #000000 !important;
+    padding: 40px 20px 20px 20px !important; /* Top padding for button space */
+    border-radius: 14px;
+    border: 1px solid rgba(255, 255, 255, 0.15);
+    overflow-x: auto;
+    margin: 25px 0 !important;
+    white-space: pre !important; 
+    tab-size: 4;
+}
+
+/* Copy Button Styling */
+.copy-btn {
+    position: absolute;
+    top: 10px;
+    right: 12px;
+    background: rgba(249, 115, 22, 0.1);
+    color: #f97316;
+    border: 1px solid rgba(249, 115, 22, 0.3);
+    padding: 4px 12px;
+    border-radius: 6px;
+    font-size: 11px;
+    cursor: pointer;
+    transition: 0.3s all ease;
+    font-family: 'Inter', sans-serif;
+    text-transform: uppercase;
+    font-weight: 600;
+    letter-spacing: 0.5px;
+}
+
+.copy-btn:hover {
+    background: #f97316;
+    color: #fff;
+}
+
+.copy-btn.copied {
+    background: #22c55e; /* Green for success */
+    border-color: #22c55e;
+    color: #fff;
+}
+
+.markdown-body code {
+    font-family: 'Fira Code', 'JetBrains Mono', monospace;
+    color: #38bdf8;
+    background: rgba(56, 189, 248, 0.1);
+    padding: 3px 7px;
+    border-radius: 6px;
+    font-size: 0.92em;
+}
+
+.markdown-body pre code {
+    background: transparent !important;
+    padding: 0 !important;
+    color: #e2e8f0 !important;
+    display: block !important;
+    line-height: 1.6;
+}
+
+/* Elite Table Styling */
+.markdown-body table {
+    width: 100%;
+    border-collapse: separate;
+    border-spacing: 0;
+    margin: 25px 0;
+    background: rgba(255, 255, 255, 0.02);
+    border: 1px solid rgba(255, 255, 255, 0.1);
+    border-radius: 12px;
+    overflow: hidden;
+}
+
+.markdown-body th {
+    background: rgba(249, 115, 22, 0.15);
+    color: #fff;
+    padding: 14px;
+    text-align: left;
+}
+
+.markdown-body td {
+    padding: 12px 14px;
+    border-top: 1px solid rgba(255, 255, 255, 0.05);
+}
+
+/* -------------------------------------------------------------------------------------*/
