+++ templates/base.html
@@ -35,6 +35,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
 
     <link rel="stylesheet" href="{% static 'css/global.css' %}" />
+    <link rel="stylesheet" href="{% static 'css/markdown-core.css' %}">
     <link rel="stylesheet" href="{% static 'fontawesome/css/all.min.css' %}" />
     <link rel="stylesheet" href="{% static 'accounts/css/auth.css' %}" />
 
@@ -45,6 +46,7 @@
     <link rel="stylesheet" href="{% static 'insights/css/insights_detail.css' %}" media="print" onload="this.media='all'" />
 
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11" defer></script>
+    <script src="{% static 'js/global-markdown.js' %}"></script>
   
 
     <style>
